fix: UI modifications
