+++ sample.env
@@ -1,0 +1,9 @@
+# This is a sample .env file for use in local development.
+# Duplicate this file as .env in the root of the project
+# and update the environment variables to match your
+# desired config
+
+REACT_APP_API_BASE_URL = https://example.com
+VITE_SERVICE_ID = randomstring
+VITE_TEMPLATE_ID = randomstring
+VITE_PUBLIC_ID = randomstring
