+++ src/APP/components/index.js
@@ -4,7 +4,6 @@ export { default as CartDrawer } from "./shop/CartDrawer";
 export { default as Counter } from "./shop/Counter";
 export { default as FAQ } from "./FAQ";
 export { default as Footer } from "./Footer";
-export { default as Footer2 } from "./Footer2";
 export { default as GoBackBtn } from "./GoBackBtn";
 export { default as Header } from "./Header";
 export { default as Loader } from "./Loader";
