+++ src/APP/index.js
@@ -14,7 +14,8 @@ import Blog2 from "./pages/blog2/Blog2";
 import Blogs from "./pages/blogs/Blogs";
 import IndividualChapter from "./pages/chapter/pages/IndividualChapter";
 import CommunityPage from "./pages/community/CommunityPage";
-import SingleEvent from "./pages/community/sections/eventsSection/SingleEvents/SingleEvent";
+// import SingleEvent from "./pages/community/sections/eventsSection/SingleEvents/SingleEvent";
+import { SingleEvent } from "./pages/community/sections";
 import DonatePage from "./pages/donate/DonatePage";
 import SingleProductDonation from "./pages/donate/pages/SingleProductDonatePage";
 import Error400 from "./pages/errorPages/Error400";
@@ -22,7 +23,7 @@ import Error403 from "./pages/errorPages/Error403";
 import Error404 from "./pages/errorPages/Error404";
 import Error500 from "./pages/errorPages/Error500";
 import ErrorBoundary from "./pages/errorPages/ErrorBoundary";
-import EventsPage from "./pages/events/pages/EventsPage2";
+import EventsPage from "./pages/events/pages/EventsPage";
 import EventsSection from "./pages/events/sections/eventsSection/EventsSection";
 import GalleryPage from "./pages/gallery/GalleryPage";
 import LandingPage from "./pages/landingPage/LandingPage";
