+++ src/APP/pages/community/sections/eventsPreview/SingleEvents/sections/index.js
@@ -1,0 +1,7 @@
+export { default as EventCard } from "./EventCard";
+export { default as EventDescription } from "./EventDescription";
+export { default as EventRSVP } from "./EventRSVP";
+export { default as Hero } from "./Hero";
+export { default as LocationTag } from "./LocationTag";
+export { default as SimilarEvents } from "./SimilarEvents";
+export { default as RSVPForm } from "./RSVPForm";
