+++ src/APP/pages/community/sections/index.js
@@ -2,7 +2,7 @@ export { default as JoinSYTModal } from "./JoinSYTModal";
 export { default as WelcomeSection } from "./WelcomeSection";
 export { default as ChaptersSection } from "./chaptersSection/ChaptersSection";
 export { default as ChapterCard } from "./chaptersSection/ChapterCard";
-export { default as SingleEvent } from "./eventsSection/SingleEvents/SingleEvent";
+export { default as SingleEvent } from "./eventsPreview/SingleEvents/SingleEvent";
 export { default as GallerySection } from "./gallerySection/GallerySection";
 export { default as PartnerCard } from "./partnerSection/PartnerCard";
 export { default as PartnerSection } from "./partnerSection/PartnerSection";
