+++ src/APP/pages/events/sections/FeaturedCarousel.jsx
@@ -29,8 +29,9 @@ export default function FeaturedCarousel() {
         </div>
       )}
     >
-      {Array.from({ length: 5 }).map((_, i) => (
-        <FeaturedEventCard key={i} />
+      {/* eslint-disable-next-line no-unused-vars */}
+      {Array.from({ length: 2 }).map((_) => (
+        <FeaturedEventCard key={crypto.randomUUID()} />
       ))}
     </Carousel>
   );
