+++ src/APP/pages/landingPage/sections/OurEvents.jsx
@@ -47,7 +47,7 @@ function OurEvents() {
           <h4 className="text-green-dark text-[32px] leading-normal font-semibold">
             Upcoming Events
           </h4>
-          <Button link="/all-events" title="view more" />
+          <Button link="/events" title="view more" />
         </div>
 
         {isError && (
